chore: inline scheduling logic
